LayoutTests/platform/efl-wk2/fast/media/mq-resolution-dpi-dpcm-warning-expected.txt

CONSOLE MESSAGE: line 9: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: screen and (min-resolution: 96dpi)
CONSOLE MESSAGE: line 9: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: (min-resolution: 160dpcm)
CONSOLE MESSAGE: line 9: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: (resolution: 96dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: (min-resolution: 2dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: (min-resolution: 2dpcm)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: screen and (min-resolution: 5dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: screen and (min-resolution: 5dpcm)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: (min-resolution: 10dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: (min-resolution: 10dpcm)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: screen and (max-resolution: 300dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: screen and (max-resolution: 300dpcm)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: (resolution: 96dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: (resolution: 96dpcm)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpi', as in CSS 'dpi' means dots-per-CSS-inch, not dots-per-physical-inch, so does not correspond to the actual 'dpi' of a screen. In media query expression: (resolution: 96dpi)
CONSOLE MESSAGE: Consider using 'dppx' units instead of 'dpcm', as in CSS 'dpcm' means dots-per-CSS-centimeter, not dots-per-physical-centimeter, so does not correspond to the actual 'dpcm' of a screen. In media query expression: (resolution: 96dpcm)
Test verifies that warning is added to console when dpi or dpcm CSS values are used for resolution media feature
